requirements.txt
feedparser
python-dotenv
requests
markdown
google-genai
firebase-admin
